+++ server.js
@@ -19,10 +19,6 @@ const server = http.createServer(app);
 const { Server } = require("socket.io");
 const io = new Server(server);
 
-let sortAlphabets = function (text) {
-  return text.split("").sort().join("");
-};
-
 app.use(express.static("static/public"));
 
 app.use(express.json());
@@ -30,37 +26,15 @@ app.use(express.urlencoded());
 app.use("/", homeRoutes);
 app.use("/chat", chatRoutes);
 
-const Conversation = require("./src/models/Conversation");
+const chatHandler = require("./chatHandler");
 
 io.on("connection", (socket) => {
   socket.on("join room", (message) => {
-    socket.join(sortAlphabets(`${message.userSelf}${message.userOther}`));
+    chatHandler.joinRoom(socket, message);
   });
 
   socket.on("chat message", (message) => {
-    Conversation.findOneAndUpdate(
-      {
-        conversationName: sortAlphabets(
-          `${message.userSelf}${message.userOther}`
-        ),
-      },
-      {
-        $push: {
-          messages: {
-            userSender: message.userSelf,
-            message: message.message,
-          },
-        },
-      },
-      {
-        upsert: true,
-      }
-    ).then(() => {
-      io.to(sortAlphabets(`${message.userSelf}${message.userOther}`)).emit(
-        "chat message",
-        { message: message.message, userSender: message.userSelf }
-      );
-    });
+    chatHandler.messagesSend(io, message);
   });
 });
